awaken api fix
