+++ src/CAServer.Application.Contracts/Tokens/ITokenNftAppService.cs
@@ -1,5 +1,6 @@
 using System.Collections.Generic;
 using System.Threading.Tasks;
+using CAServer.Awaken;
 using CAServer.Tokens.Dtos;
 using CAServer.UserAssets;
 using CAServer.UserAssets.Dtos;
@@ -12,4 +13,7 @@ public interface ITokenNftAppService
     Task<List<GetTokenListDto>> GetTokenListAsync(GetTokenListRequestDto input);
     Task<SearchUserPackageAssetsDto> SearchUserPackageAssetsAsync(SearchUserPackageAssetsRequestDto requestDto);
     Task<SearchUserAssetsDto> SearchUserAssetsAsync(SearchUserAssetsRequestDto requestDto);
+    
+    Task<AwakenSupportedTokenResponse> ListAwakenSupportedTokensAsync(int skipCount, int maxResultCount,
+        int page, string chainId, string caAddress);
 }
